test: Added test to highlight terminal rendering issue.
This change adds a test that is expected to fail in batch mode and pass
in normal mode.

Terminal display is broken when run in batch mode: the lines of a
window to the right of a buffer showing a line with no final \n
on it are truncated, so the positions don't align.

So the terminal would look like this:

 mostly empty       |line 0------------$|mostly empty
                    |line 1------------$|
 . |line 2------------$|.
 |line 3------------$|
 |line 4------------$|
 |line 5------------$|
 |line 6------------$|
 |line 7------------$|
 [...]

This is puzzling and very annoying as tests are primarily run in batch
mode.