feat: Detect specific faces from color and recover them.
With this change, it becomes possible to give a list of faces to be
copied and find the sections displayed with these faces marked with a
'face text property.

Internally, termgrab assigns special color to these faces that stand
out, just before grabbing, then maps these colors back to the original
faces.