fix: Fix termgrab-with-grab-buffer and panel resize.
Before this change, termgrab-with-grab-buffer wouldn't execute its body,
which made many tests just appear to pass.

Notable tests that appeared to pass where tests for termgrab-resize.
Turned out that termgrab-resize didn't work reliably, as resizing a pane
can take a long and unpredictable time. This change restarts emacsclient
to be sure the terminal changed.

This change also changes the API of terminal-resize. Instead of a
separate function, the size is now part of the signature of
start-server.