feat: termgrab-read-from-minibuffer - BROKEN in batch mode.
This change introduces termgrab-read-from-minibuffer which allows
grabbing the screen while reading from the minibuffer.

Grabbing the screen cannot happen as long as there's pending input,
because Emacs won't redisplay, so the approach taken by ert won't allow
grabbing, as read-from-minibuffer runs only as long as there are keys to
read.

But in batch mode, read-from-minibuffer reads from stdin, unless it is
in a macro. So the approach in this change doesn't work.