fix: Accept nil as :result in a response.
Before this change, when dispatching a message, the code would
use (plist-get msg :result) to check whether there was a result, but
that failed if the result was nil. With this change, the code
uses (plist-membor msg :result), which tests for the presence of
:result.

Other members of the response, such as :id or :error still must be
non-nil.

This change also adds some commented-out messages useful for debugging
such issues, waiting for some kind of debug logging facility.