feat: Treat errors in :error as Emacs error symbol + data.
Before this change, errors response could contain anything. They would
just be used to create an error with a string containing the original
error. This wasn't convenient when the remote error symbol had meaning.

With this change, the error response is treated as a real Emacs error,
which can be caught using condition-case.