fix: Fix and test face recognition.
Before this change, faces to be grabbed were not recognized, because the
screen was re-displayed too early. This change fixes that and adds test
cases for face recognition.

Before this change, ansi colors were used for face recognition, but that
was just confusing and dangerous, as ansi colors are customizable.