fix: Make sure unreadable types don't break the whole message.
Before this change, unreadable types were just skipped, which would
sometimes make the whole message unreadable. That wasn't the intent.

With this change, unreadable types are serialized as (unreadable
typename), which, while not very useful, will at least not break the
whole message containing an unreadable type.

This is still useful as often forms end with a function that returns
some unreadable type that'll just be ignored on the other end, treating
the function as void.