feat: Have the client send its messages to the server during eval.
Before this change, getting the messages logged during a turtles test
required checking the messages on the client instance, which might not
be possible in batch mode.

With this change, client messages are available in the *Messages* of the
server, in interactive mode, or are printed to stdout, in batch mode.