feat: Support Emacs 26.1 and later.
Before this change, Turtles require 29.1. With this cange, Turtles works
on Emacs 26.1 and later, with the help of a new dependency on compat.

Limitations:

- term.el didn't support 24bit colors before Emacs 29. Earlier versions
are very limited when it comes to colors (16 in Emacs 28, 8 before
that).

- clear-minibuffer-message isn't available before Emacs 29, so messages
may be left between tests

- condition-case doesn't support t as a condition to catch everything.
We can only catch errors on older Emacs versions.

Issues that could be worked around:

- compat provides support for ert-with-temp-directory and others

- print-unreadable-function isn't available until Emacs 29.1, so this
version is using search-and-replace on the output of print

- split-window-below works only on the selected window

- pcase-let isn't available

- turtles-display-buffer-full-frame takes the place of Emacs 29.1
display-buffer-frame

- ansi colors aren't available for tests, using hardcoded hex colors

- the struct ert-test has no file-name, capturing the filename in
turtles-ert-test, now a macro

- subr-x must be required to use when-let

- set-message-function isn't available until Emacs 29, using advices
instead