fix: Handle messages from possibly concurrent 'eval better.
Before this change, an advice was added before handling the method 'eval
in the client then removed afterwards. This only works if the calls are
always in sequence, but things can get a little strange when working
with minibuffer-read.

This change switches to keeping the advince always set, and using a
variable to figure out whether to send messages up.

This change also adds a variable to avoid sending messages while
handling messages, which would cause an infinite loop.