fix: Fix code parsing unreadable types from prin1.
Before this change, unreadables were matched with the regexp "#<.*?>"",
which assumed the text within #<> didn't contain any >, but that's not
the case. Buffer, frames, processes and others can contain anything, as
their name is printed without any escaping.

This change matches unreadables found in the output of prin1 with the
longest known name for the type, as a workaround, then generates type
specific object representation:

  (turtles-buffer :name ...)
  (turtles-frame :name ...)
  (turtles-process :name ...)
  (turtles-window :buffer ...)
  (turtles-overlay :from ... :to ... :buffer)
  (turtles-marker :pos ... :buffer)
  (turtles-object ...)