refactor: Have instances register themselves to the server.
Before this change, instances were detected when they connected to the
server, but they might not have been ready yet to accept method calls.

With this change, instances are only used after they register themselves
to the server. Registration also gives a change for instances to
identify themselves, in follow-up changes.