feat: Load source file only once per test suite.
Before this change, when running a set of tests with ert-run-tests, each
test would load its file on the instance. This slowed things down
noticeably.

This change advises ert-run-tests to allow turtles-ert to track which
files have been loaded and so only load it the first time on a given
instance.