feat: Try to run tests without file-name.
Before this change, turtles-ert-test would fail when a test had no
file-name. This tended to happen when re-running tests.

With this change, if turtles-ert-test doesn't have a file-name, it sends
the whole test to the instance, including the test body, and execute
that. This might fail if the test body references something that's not
available, but it will normally work just fine when re-running tests.