refactor: Add eat as an optional test dependency
In this change, the point at which turtles interfaces with the terminal
are extracted into generic functions in turtles-term.el, implemented for
term in turtles-term-term.el and for eat in turtles-term-eat.el.

With the new interface, screen grabs are new defined as storing face
information into the face property instead of the font-lock-face
property.