refactor: Work with 'face instead of 'font-lock-face.
Before this change, turtles.el expected screen grabs to have face
information in font-lock-face properties, which is the way term.el does
it.

However, with the switch to supporting eat, terminals should now all set
the face property only and not font-lock-face.

This change adapts existing code to this interface change.