feat: Define and load "terminal extensions" dynamically.
With this change, turtles-instances loads extensions by prepending the
turtles-instance-terminal with "turtles-", so that gives us
"turtles-term" and "turtles-eat". This way, it's truly "extensible".

This change also defines turtles-eat as an extension in Eldev and in the
source tree: turtles-eat.el is moved into the extensions subdirectory,
defined as a single-file package that depend on turtles and excluded
from the package.

extensions/turtles-eat.el is however compiled and included in the tests.