fix: Include terminal content when Emacs fails to connect.
When Emacs fails to connect, usually it means that it failed to start.
To help debug that, the error message included the term buffer, but:
- the buffer might be too large to be readable
- the buffer was included as a string, notably the newlines were
escaped. This made it hard to read.

This change includes only the first 60 chars of the buffer directly into
the message (no format) and cleans up the code.