feat: turtles-read-from-minibuffer exits implicitly.
Before this change, it was necessary to exit the minibuffer at the end
of the body of turtles-read-from-minibuffer or it'd throw an error. The
result was that all calls of turtles-read-from-minibuffer ended
with (exit-minibuffer).

This change adds an implicit (exit-minibuffer) at the end of the BODY of
turtles-read-from-minibuffer. It is still possible and supported to
abort the minibuffer in other ways, but if the minibuffer is still
active at the end of the body, it is simply exited that way.