feat: Support :keys in turtles-read-from-minibuffer.
This change adds the possibility of feeding keys to the Emacs instance
to be processed by the normal command loop from within
turtles-read-from-minibuffer. This avoids the shortcomings that come
from the various ways can be simulated, using execute-kbd-macro or
ert-simulate-keys.

The way this is implemented, the instance asks the upstream Emacs to
send it a terminal sequence that corresponds to F62 in some terminals
for each key to be pressed, then translate that key into key events
coming from an internal stack using key-translation-map. This avoids
having to convert key events into something the terminal will understand
and back.