feat: Support :command in turtles-read-from-minibuffer.
With this change, it becomes possible to execute commands as if from the
normal command loop in turtles-read-from-minibuffer without having to
bind them first.

In the background, the command is bound to a transient map, so for one
call only. What it is bound to can be configured, if necessary, using
:command-with-keybinding.