fix: Fix window grabbing with larger term sizes.
Before this change, basic window grabbing would fail for any term-based
terminal that wasn't 80x20. This was because term would ignore the
initial width and height when drawing, but still keep term-width and
term-height set to these values. The result was inconsistent.

With this change, terminals are always created with their default size,
then resized later on if another size is specified.