feat: Set default terminal size to 80x24, define a larger 132x43.
Before this change, the default terminal size was 80x20, inherited from
tmux. This change switches to a more common 80x24.

This change also defines a larger terminal size and tests both terminal
size and their corresponding frame size.

Additionally, the call to switch off menu-bar mode is moved to
turtles--launch, so that term-height = frame-height from the very
beginning. Having frame-height change because the menu was not disabled
immediately was much too confusing.