fix: Workaround for issue with pos before a display replacement.
Before this change, the position just before a display replacement was
not available when in a minibuffer window. Position would immediately
skip to after the replacement.

The behavior in a normal buffer was different, and more reasonable: the
positions both before and after the display text replacement are
accessible.

This change works around this specific issue and adds test for both the
normal and the minibuffer cases.