fix: Take cursor position from terminal.
Before this change, point was recovered in a window by converting a
buffer position to screen coordinates. That wasn't reliable in a
minibuffer, when using the display text property.

This change switches to another approach, which is to select the window
that is about to be grabbed, then move the point to where the cursor is
on the terminal. This is simpler and should be more reliable.

This change removes:

 - grabbing mark and region, since this cannot be done using this
 approach

 - turtles-pos-in-window-grab, since it wasn't reliable

 - turtles-source-window and turtles-source-buffer, since they're
 not useful anymore.