feat: Forward variable values from server to instance.
This test adds the option :forward to turtles-definstance that allows
defining a set of variable whose value must be forwarded to the instance
when they're started.

This is especially useful for batch tests, where options can be set
globally in the test environment to configure them. :forward can then
forward the value of these option to the instances.