refactor: Change turtles-io-method-handler to not return a lambda.
Before this change, most methods handlers were defined as a lambda,
using turtles-io-method-handler. This had a big disadvantage: reloading
the file would not change the way the method was handled. Only a restart
would do.

With this change, most method handlers are defined as a function that
then calls turtles-io-handle-method. Such functions can be reloaded.