feat!: Replace :keys, etc. with (turtles-input-keys) etc.
Before this change, it was possible to feed events through the command
loop from turtles-with-minibuffer with :events :keys :command. This was
limited to the toplevel flow of the body of turtles-with-minibuffer.

This change turns these special constructs into functions:
 (turtles-input-keys keys)
 (turtles-input-events events)
 (turtles-input-command command &optional keybinding)

These functions can be called from outside turtles-with-minibuffer. They
work more by opening a new (recursive-edit).