feat: Go back to :keys, etc. instead of (turtles-input-keys) etc.
The implementation of turtles-input-keys, based on recursive-edit,
turned out to cause too many issues with minibuffer interactions, the
very thing that needs that functionality.

This change goes back to the previous implementation, which define :keys
etc. as special form in turtles-with-minibuffer. These don't need
recursive edits and so are closer to the real thing.