feat: Improve pop-to-buffer selection UI.
Before this change, the user was supposed to select the short
documentation of the functions, which might not be very practical,
depending on how the user's completing-read works.

With this change, the user is supposed to select a short symbol, with
the shortdoc displayed as annotation.

This change also adds tests for the pop-to-buffer selection UI.