feat!: Introduce turtles-ert-deftest. Replaces turtles-ert-test.
Before this change, turtle tests were written:

(ert-deftest name ()
  ...
  (turtles-ert-test)
  ...test)

With this change, turtle tests become:

(turtles-ert-deftest name ()
  ...test)

This new approach has the following advantages:

We don't have to register advices during load. They're now loaded by
turtles-ert-deftest and only on the server

We don't have to worry about (ert-test-file). The file tracking that
turtles-ert-deftest does now is just as effective.

There's no need to worry about a part, before (ert-test-file), running
on the server; everything is run on the instance.

Having a different deftest call probably also makes it clearer that this
is a special test.

This is a backward-incompatible change, so the version is bumped up.