fix: Synchronize ert-test timeout with minibuffer timeouts.
With this change, the timeout used by turtles-with-minibuffer is set
just one second short of the server-side timeout set for the test
itself.

This change allows providing a more useful error message than "ert test
timed out" in the case where the BODY section of turtles-with-minibuffer
failed to make the READ section return properly.