fix: Set test timeout when running tests without files.
Before this change, turtles--ert-test-abs-timeout was only set when
running tests from file.

This change unifies the test-running logic in both cases (with and
without file) to fix the problem and avoid such issues in the future. It
also tests the shorter timeout set in turtles--ert-test-abs-timeout and
complains if it's too short.